Clarified when number of team leads can be reduced
`2.6` provided for increasing the number of team leads, but it only allowed reducing it when a team lead resigns or ceases to participate. Expanded the language to allow the Steering Committee to also be able to reduce it at the end of the team lead's term.

Passed by voice vote during the 2023-01-24 Steering Committee meeting as described in the minutes.